requirements.txt
pipecat-ai[livekit,openai,deepgram,silero,flows] 
pipecat-ai-flows
python-dotenv 
loguru

fastapi
